[refactor] 내가 저장한 이벤트 가져오는 함수명 수정 (#93)
