+++ src/main/java/hyangyu/server/api/MyPageApi.java
@@ -1,8 +1,8 @@
 package hyangyu.server.api;
 
 import hyangyu.server.dto.ErrorDto;
-import hyangyu.server.dto.MyPageDto;
-import hyangyu.server.dto.MyPageResponseDto;
+import hyangyu.server.dto.mypage.MyPageDto;
+import hyangyu.server.dto.mypage.MyPageResponseDto;
 import hyangyu.server.dto.UserDto;
 import hyangyu.server.dto.event.*;
 import hyangyu.server.service.FavoriteDisplayService;
