+++ src/main/java/hyangyu/server/dto/mypage/MyPageDto.java
@@ -1,11 +1,8 @@
 package hyangyu.server.dto;
 
-import hyangyu.server.dto.EventDto;
 import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
-
-import java.util.List;
 
 @Data
 @AllArgsConstructor
